changed / fixed inconsistency / minor optimization in Index.union when one contained in the other, fixed resulting failing unit tests. forced SparseSeries fill_value to always be float64 type

+++ pandas/core/sparse.py
@@ -239,7 +239,7 @@ class SparseSeries(Series):
         output = subarr.view(cls)
         output._sp_values = subarr
         output.sp_index = sparse_index
-        output.fill_value = fill_value
+        output.fill_value = np.float64(fill_value)
         output.index = index
         return output
